Move interfaces to seperate type.ts
