+++ src/plays/quiz-app/types.ts
@@ -1,0 +1,12 @@
+interface Status {
+  qNo: number;
+  status?: number;
+}
+
+interface QuestionProps {
+  qNo: number;
+  qStatus: number;
+  question: string;
+}
+
+export { Status, QuestionProps }
